requirements.txt
pybit==5.8.0
pandas==2.2.0
numpy==1.26.4
ta==0.11.0
flask==3.0.2
flask-socketio==5.3.6
python-dotenv==1.0.1
requests==2.31.0
apscheduler==3.10.4
eventlet==0.36.1